fix(AutoPlay): Fixed issue with auto-play not working on /user/.
Fixed issue with auto-play not working on /user/ and if going to the profile page twice.
Fixed issue with the detail page player not auto-playing even if auto-play is disabled by calling `player.play()`.
Fixed issue with auto-play patch being applied twice.
Fixed issue with an error caused the connection to be disposed of.
Fixed issue with calling destroy on the player would cause the registered player to get disposed of. Apparently a destroyed player can be reused (memory leak?).

Closes #52, #17